refactor: test layer와 production layer 분리 위해 test 폴더 도입
