+++ docs/REFACTORING.md
@@ -84,7 +84,7 @@
 
 3.  테스트
 
-    :.: 테스트를 위한 코드는 프로덕션 코드에서 분리
+    :V: 테스트를 위한 코드는 프로덕션 코드에서 분리
 
     :.: 단위 테스트 실패 이유는 하나!
 
@@ -171,3 +171,5 @@
     :.: test layer와 production layer 분리 의미 = testFixures 어디에 위치?
 
     :.: 테스트가 필요한 핵심적인 private 코드가 있다면 역할 분리 고민 -> GenerateNumber 분리 ?
+
+    :.: test용 파일을 위한 test 코드 -> 배보다 배꼽이 커지는 건 아닌지?
