+++ test/FixedStrategy.js
